improve folder with utils, types and schemas
