+++ packages/swaig/src/index.ts
@@ -1,1 +1,3 @@
 export { SWAIG } from './swaig'
+
+export * from './types'
